Сделал рандомную генерацию времени
